requirements.txt
pymongo==4.7.2
requests==2.31.0
openai==1.30.1
python-dotenv==1.0.1
